#273 Fix stacktrace when Marathon app already exists
Also removed MarathonClient and moved code to Marathon.
