+++ test-framework-docker/system-test/src/test/java/com/containersolutions/mesoshelloworld/systemtest/InstallCommandTest.java
@@ -3,13 +3,11 @@ package com.containersolutions.mesoshelloworld.systemtest;
 import com.containersol.minimesos.cluster.MesosCluster;
 import com.containersol.minimesos.docker.DockerContainersUtil;
 import com.containersol.minimesos.marathon.Marathon;
-import com.containersol.minimesos.marathon.MarathonClient;
 import com.containersol.minimesos.mesos.ClusterArchitecture;
 import com.containersol.minimesos.mesos.DockerClientFactory;
 import com.containersolutions.mesoshelloworld.executor.Executor;
 import com.containersolutions.mesoshelloworld.scheduler.Configuration;
 import com.github.dockerjava.api.DockerClient;
-import com.jayway.awaitility.Awaitility;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import org.apache.commons.io.IOUtils;
@@ -24,6 +22,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
+import static com.jayway.awaitility.Awaitility.*;
 import static org.junit.Assert.fail;
 
 public class InstallCommandTest {
@@ -43,27 +42,11 @@ public class InstallCommandTest {
 
     @Test
     public void testMesosInstall() throws IOException {
-
-        File taskFile = new File( "src/test/resources/test-framework-docker.json" );
-        if (!taskFile.exists()) {
-            fail("Failed to find task info file " + taskFile.getAbsolutePath());
-        }
-
-        // wait for Marathon to start
-        MarathonClient marathon = new MarathonClient( CLUSTER.getMarathonContainer().getIpAddress() );
-        Awaitility.await("Marathon did not start responding").atMost(60, TimeUnit.SECONDS).until(marathon::isReady);
-
-        String masterIp = CLUSTER.getMasterContainer().getIpAddress();
-
-        try (FileInputStream fis = new FileInputStream(taskFile)) {
-            String appJson = IOUtils.toString(fis);
-            appJson = appJson.replace(MESOS_MASTER_IP_TOKEN, masterIp);
-            CLUSTER.getMarathonContainer().deployApp(appJson);
-        }
+        deployApp();
 
         DockerContainersUtil util = new DockerContainersUtil(CONFIG.dockerClient);
         final List<String> ipAddresses = new ArrayList<>();
-        Awaitility.await("executors are expected to come up").atMost(60, TimeUnit.SECONDS).until(() -> {
+        await("executors are expected to come up").atMost(60, TimeUnit.SECONDS).until(() -> {
             ipAddresses.clear();
             ipAddresses.addAll(util.getContainers(false).filterByImage(Configuration.DEFAULT_EXECUTOR_IMAGE).getIpAddresses());
             return ipAddresses.size() > 0;
@@ -71,8 +54,7 @@ public class InstallCommandTest {
 
         String appUrl1 = "http://" + ipAddresses.get(0) + ":8081";
         String appUrl2 = "http://" + ipAddresses.get(0) + ":8082";
-        Awaitility.await("The app did not start providing expected output").atMost(60, TimeUnit.SECONDS).until( () -> {
-
+        await("The app did not start providing expected output").atMost(60, TimeUnit.SECONDS).until( () -> {
             String response1;
             try {
                 response1 = Unirest.get(appUrl1).asString().getBody();
@@ -86,22 +68,31 @@ public class InstallCommandTest {
                 response2 = "";
             }
 
-            return response1.equals( Executor.RESPONSE_STRING ) || response2.equals( Executor.RESPONSE_STRING );
-
+            return response1.equals(Executor.RESPONSE_STRING) || response2.equals(Executor.RESPONSE_STRING);
         });
+    }
 
+    private void deployApp() throws IOException {
+        File taskFile = new File("src/test/resources/test-framework-docker.json");
+        if (!taskFile.exists()) {
+            fail("Failed to find task info file " + taskFile.getAbsolutePath());
+        }
+
+        try (FileInputStream fis = new FileInputStream(taskFile)) {
+            String appJson = IOUtils.toString(fis);
+            appJson = appJson.replace(MESOS_MASTER_IP_TOKEN, CLUSTER.getMasterContainer().getIpAddress());
+            CLUSTER.getMarathonContainer().deployApp(appJson);
+        }
     }
 
     @AfterClass
     public static void removeExecutors() {
-
         DockerContainersUtil util = new DockerContainersUtil(CONFIG.dockerClient);
 
         // stop container, otherwise it keeps on scheduling new executors as soon as they are stopped
         util.getContainers(false).filterByImage(SchedulerContainer.SCHEDULER_IMAGE).kill().remove();
         // remove executors
         util.getContainers(false).filterByImage(Configuration.DEFAULT_EXECUTOR_IMAGE).kill().remove();
-
     }
 
 }
